requirements.txt
api==0.0.7
argcomplete==2.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.17
botocore==1.27.17
certifi==2022.6.15
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-filter==22.1
django-rest-framework==0.1.0
django-s3-storage==0.13.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
nose==1.3.7
packaging==21.3
placebo==0.9.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.0
troposphere==4.0.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wsgi-request-logger==0.4.6
zappa==0.54.1
